[level 1] Title: 문자열을 정수로 바꾸기, Time: 0.02 ms, Memory: 61.9 MB -BaekjoonHub

+++ 프로그래머스/lv1/12925. 문자열을 정수로 바꾸기/문자열을 정수로 바꾸기.kt
@@ -1,5 +1,5 @@
 class Solution {
-    fun solution(s: String): Long {
-       return s.toLong()
+    fun solution(s: String): Int {
+       return s.toInt()
     }
 }
